WEAT-9 impl network logic for weekly weather

+++ core/domain/src/main/java/io/wookoo/domain/model/weather/current/PrecipitationModel.kt
@@ -1,7 +1,7 @@
 package io.wookoo.domain.model.weather.current
 data class PrecipitationModel(
-    val level: Float,
-    val rain: Float,
-    val showers: Float,
-    val snowfall: Float,
+    val level: Double,
+    val rain: Double,
+    val showers: Double,
+    val snowfall: Double,
 )
